% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ip4}
\alias{as.character.ip4}
\alias{as.ip4}
\alias{as.ip4.character}
\alias{as.ip4.numeric}
\alias{ip4}
\title{IPv4 Conversion Functions}
\usage{
as.ip4(x)

\method{as.ip4}{numeric}(x)

\method{as.ip4}{character}(x)

\method{as.character}{ip4}(x, ...)
}
\arguments{
\item{x}{An object to be converted.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Functions to manipulate objects of class \code{"ip4"} representing IP version 4 addresses.
}
\examples{
as.ip4("192.168.0.1")
}

