% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ip6}
\alias{as.character.ip6}
\alias{as.ip6}
\alias{as.ip6.character}
\alias{ip6}
\title{IPv6 Conversion Functions}
\usage{
as.ip6(x)

\method{as.ip6}{character}(x)

\method{as.character}{ip6}(x, ...)
}
\arguments{
\item{x}{An object to be converted.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Functions to manipulate objects of class \code{"ip6"} representing IP version 6 addresses.
}
\examples{
as.ip6("DE:AD:BE:EF:01:23:45:67")
}

