% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mySqlToIp4}
\alias{mySqlToIp4}
\title{Convert MySQL ATONs to Rip4}
\usage{
mySqlToIp4(x)
}
\arguments{
\item{x}{input numeric vector}
}
\value{
integer format IP addresses
}
\description{
When you query a MySQL db, you should use INET_ATON(ip_field) to convert a string IP
to an (unsigned) int.
}
\details{
Unfortunately, the RMySQL driver will convert these to doubles because not all
unsigned ints are representable as ints. Because doubles get normalized, we
can't mask subnets directly.

Instead, we can map the top half of unsigned integers to the negative half of signed
ints using 2s-complement.
}

