% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{classA}
\alias{classA}
\alias{classB}
\alias{classC}
\title{Extract Networks from IP addresses}
\usage{
classA(x)

classB(x)

classC(x)
}
\arguments{
\item{x}{an \code{ip4} object}
}
\description{
Extract Networks from IP addresses
}

