% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mask}
\alias{classA}
\alias{classB}
\alias{classC}
\alias{mask}
\title{Extract Networks from IP addresses}
\usage{
mask(x, m)

classA(x)

classB(x)

classC(x)
}
\arguments{
\item{x}{an \code{ip4} object}

\item{m}{a bit mask}
}
\description{
Extract Networks from IP addresses
}

